Improve I
Still doesn't pass: time limit exceeded.

+++ lab1/I.cpp
@@ -12,15 +12,15 @@ int main(){
 		std::cin >> n;
 		std::cin >> k;
 
-		int a = k;
-		int b = std::stoi(std::to_string(k * k).substr(0, n));
-		int max = std::max(a, b);
+		long long int a = k;
+		long long int b = std::stoll(std::to_string(k * k).substr(0, n));
+		long long int max = std::max(a, b);
 
 		while(a != b){
-			a = std::stoi(std::to_string(a * a).substr(0, n));
-			b = std::stoi(std::to_string(b * b).substr(0, n));
-			b = std::stoi(std::to_string(b * b).substr(0, n));
-			int temp = std::max(a, b);
+			a = std::stoll(std::to_string(a * a).substr(0, n));
+			b = std::stoll(std::to_string(b * b).substr(0, n));
+			b = std::stoll(std::to_string(b * b).substr(0, n));
+			long long int temp = std::max(a, b);
 			if(temp > max){
 				max = temp;
 			}
